refac : 멤버 서비스 세분화
