+++ src/main/java/atwoz/atwoz/member/presentation/MemberController.java
@@ -4,7 +4,8 @@ import atwoz.atwoz.auth.presentation.AuthContext;
 import atwoz.atwoz.auth.presentation.AuthPrincipal;
 import atwoz.atwoz.common.enums.StatusType;
 import atwoz.atwoz.common.response.BaseResponse;
-import atwoz.atwoz.member.application.MemberService;
+import atwoz.atwoz.member.application.MemberContactService;
+import atwoz.atwoz.member.application.MemberProfileService;
 import atwoz.atwoz.member.application.dto.MemberContactResponse;
 import atwoz.atwoz.member.application.dto.MemberProfileResponse;
 import atwoz.atwoz.member.application.dto.MemberProfileUpdateRequest;
@@ -16,41 +17,42 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/member")
 @RequiredArgsConstructor
 public class MemberController {
-    private final MemberService memberService;
+    private final MemberContactService memberContactService;
+    private final MemberProfileService memberProfileService;
 
     @PutMapping("/profile")
     public ResponseEntity<BaseResponse<MemberProfileResponse>> updateProfile(@RequestBody MemberProfileUpdateRequest request, @AuthPrincipal AuthContext authContext) {
-        MemberProfileResponse response = memberService.updateMember(authContext.getId(), request);
+        MemberProfileResponse response = memberProfileService.updateMember(authContext.getId(), request);
         return ResponseEntity.ok(BaseResponse.of(StatusType.OK, response));
     }
 
     @GetMapping("/profile")
     public ResponseEntity<BaseResponse<MemberProfileResponse>> getMyProfile(@AuthPrincipal AuthContext authContext) {
-        MemberProfileResponse response = memberService.getProfile(authContext.getId());
+        MemberProfileResponse response = memberProfileService.getProfile(authContext.getId());
         return ResponseEntity.ok(BaseResponse.of(StatusType.OK, response));
     }
 
     @PostMapping("/profile/dormant")
-    public ResponseEntity<BaseResponse<Void>> transitionToDormant(@AuthPrincipal AuthContext authContext) {
-        memberService.transitionToDormant(authContext.getId());
+    public ResponseEntity<BaseResponse<Void>> changeToDormant(@AuthPrincipal AuthContext authContext) {
+        memberProfileService.changeToDormant(authContext.getId());
         return ResponseEntity.ok(BaseResponse.from(StatusType.OK));
     }
 
     @GetMapping("/profile/contact")
     public ResponseEntity<BaseResponse<MemberContactResponse>> getMyContact(@AuthPrincipal AuthContext authContext) {
-        MemberContactResponse response = memberService.getContacts(authContext.getId());
+        MemberContactResponse response = memberContactService.getContacts(authContext.getId());
         return ResponseEntity.ok(BaseResponse.of(StatusType.OK, response));
     }
 
     @PatchMapping("/profile/contact/kakao")
     public ResponseEntity<BaseResponse<Void>> updateKakaoId(@AuthPrincipal AuthContext authContext, @RequestBody String kakaoId) {
-        memberService.updateKakaoId(authContext.getId(), kakaoId);
+        memberContactService.updateKakaoId(authContext.getId(), kakaoId);
         return ResponseEntity.ok(BaseResponse.from(StatusType.OK));
     }
 
     @PatchMapping("/profile/contact/phone")
     public ResponseEntity<BaseResponse<Void>> updatePhoneNumber(@AuthPrincipal AuthContext authContext, @RequestBody String phone) {
-        memberService.updatePhoneNumber(authContext.getId(), phone);
+        memberContactService.updatePhoneNumber(authContext.getId(), phone);
         return ResponseEntity.ok(BaseResponse.from(StatusType.OK));
     }
 }
